moved afk to a new file
